[feat] add useCustomToast
